1013 퀴즈 JSON으로 불러와서 db에 저장. db에서 퀴즈 아이디1 조회해서 퀴즈 로드 후 화면에 표시, 참가자 목록 화면에 표시, 채팅창 구현 완료.
해야할것

퀴즈 음악처리, 퀴즈방CSS, 정답부분, 방생성시 최대인원 및 최대곡수 로직, 퀴즈 리워드 부분

+++ src/main/java/kr/co/duck/repository/QuizRoomRepository.java
@@ -13,7 +13,7 @@ import org.springframework.data.jpa.repository.Query;
 import kr.co.duck.domain.QuizRoom;
 
 // 기능: 퀴즈방 레포지토리
-public interface QuizRoomRepository extends JpaRepository<QuizRoom, Long> {
+public interface QuizRoomRepository extends JpaRepository<QuizRoom, Integer> {
 
 	// 퀴즈방 전체 조회 페이징 처리
 	Page<QuizRoom> findAll(Pageable pageable);
@@ -22,10 +22,10 @@ public interface QuizRoomRepository extends JpaRepository<QuizRoom, Long> {
 	Page<QuizRoom> findByQuizRoomNameContaining(Pageable pageable, String keyword);
 
 	// 퀴즈방 단건 조회
-	Optional<QuizRoom> findByQuizRoomId(Long quizRoomId);
+	Optional<QuizRoom> findByQuizRoomId(int quizRoomId);
 
 	// 비관적 락 적용을 통한 퀴즈방 단건 조회
 	@Lock(LockModeType.PESSIMISTIC_WRITE)
 	@Query("select q from QuizRoom q where q.quizRoomId = :quizRoomId")
-	Optional<QuizRoom> findByQuizRoomIdWithLock(Long quizRoomId);
+	Optional<QuizRoom> findByQuizRoomIdWithLock(int quizRoomId);
 }
